Crate datamodel_connector

Modules

connector_error

helper

Structs

ConnectorCapabilities

Contains all capabilities that the connector is able to serve.

EmptyDatamodelConnector

A Connector implementor meant to be used as a default when no datasource is defined.

Enums

ConnectorCapability

ConstraintScope

A scope where a constraint name must be unique.

ConstraintType

ReferentialIntegrity

Defines the part of the stack where referential actions are handled.

Traits

Connector