Trait datamodel_connector::Connector

pub trait Connector: Send + Sync {

    fn name(&self) -> &str;

    fn capabilities(&self) -> &'static [ConnectorCapability];

    fn constraint_name_length(&self) -> usize;

    fn referential_actions(
        &self, 
        referential_integrity: &ReferentialIntegrity
    ) -> BitFlags<ReferentialAction>;

    fn available_native_type_constructors(
        &self
    ) -> &'static [NativeTypeConstructor];

    fn scalar_type_for_native_type(&self, native_type: Value) -> ScalarType;

    fn default_native_type_for_scalar_type(
        &self, 
        scalar_type: &ScalarType
    ) -> Value;

    fn native_type_is_default_for_scalar_type(
        &self, 
        native_type: Value, 
        scalar_type: &ScalarType
    ) -> bool;

    fn parse_native_type(
        &self, 
        name: &str, 
        args: Vec<String>
    ) -> Result<NativeTypeInstance, ConnectorError>;

    fn introspect_native_type(
        &self, 
        native_type: Value
    ) -> Result<NativeTypeInstance, ConnectorError>;

    fn validate_url(&self, url: &str) -> Result<(), String>;

    fn allowed_referential_integrity_settings(
        &self
    ) -> BitFlags<ReferentialIntegrity> { ... }

    fn default_referential_integrity(&self) -> ReferentialIntegrity { ... }

    fn has_capability(&self, capability: ConnectorCapability) -> bool { ... }

    fn supports_composite_types(&self) -> bool { ... }

    fn supports_named_primary_keys(&self) -> bool { ... }

    fn supports_named_foreign_keys(&self) -> bool { ... }

    fn supports_named_default_values(&self) -> bool { ... }

    fn supports_referential_action(
        &self, 
        integrity: &ReferentialIntegrity, 
        action: ReferentialAction
    ) -> bool { ... }

    fn validate_field_default_without_native_type(
        &self, 
        _field_name: &str, 
        _scalar_type: &ScalarType, 
        _default: Option<&DefaultValue>, 
        _errors: &mut Vec<ConnectorError>
    ) { ... }

    fn validate_native_type_arguments(
        &self, 
        _native_type: &NativeTypeInstance, 
        _scalar_type: &ScalarType, 
        _: &mut Vec<ConnectorError>
    ) { ... }

    fn validate_model(&self, _: &Model, _: &mut Vec<ConnectorError>) { ... }

    fn constraint_violation_scopes(&self) -> &'static [ConstraintScope] { ... }

    fn find_native_type_constructor(
        &self, 
        name: &str
    ) -> Option<&NativeTypeConstructor> { ... }

    fn render_native_type(&self, native_type: Value) -> String { ... }

    fn set_config_dir<'a>(
        &self, 
        config_dir: &Path, 
        url: &'a str
    ) -> Cow<'a, str> { ... }

    fn supports_scalar_lists(&self) -> bool { ... }

    fn supports_relations_over_non_unique_criteria(&self) -> bool { ... }

    fn supports_enums(&self) -> bool { ... }

    fn supports_json(&self) -> bool { ... }

    fn supports_auto_increment(&self) -> bool { ... }

    fn supports_non_id_auto_increment(&self) -> bool { ... }

    fn supports_multiple_auto_increment(&self) -> bool { ... }

    fn supports_non_indexed_auto_increment(&self) -> bool { ... }

    fn supports_compound_ids(&self) -> bool { ... }

    fn allows_relation_fields_in_arbitrary_order(&self) -> bool { ... }

    fn native_instance_error(
        &self, 
        instance: &NativeTypeInstance
    ) -> ConnectorErrorFactory { ... }

    fn native_str_error(&self, native_str: &str) -> ConnectorErrorFactory { ... }

    fn native_types_not_supported(
        &self
    ) -> Result<NativeTypeInstance, ConnectorError> { ... }

}

Required methods

fn name(&self) -> &str

fn capabilities(&self) -> &'static [ConnectorCapability]

fn constraint_name_length(&self) -> usize

The maximum length of constraint names in bytes. Connectors without a limit should return usize::MAX.

fn referential_actions(
&self,
referential_integrity: &ReferentialIntegrity
) -> BitFlags<ReferentialAction>

fn available_native_type_constructors(&self) -> &'static [NativeTypeConstructor]

Returns all available native type constructors available through this connector. Powers the auto completion of the VSCode plugin.

fn scalar_type_for_native_type(&self, native_type: Value) -> ScalarType

Returns the Scalar Type for the given native type

fn default_native_type_for_scalar_type(&self, scalar_type: &ScalarType) -> Value

On each connector, each built-in Prisma scalar type (Boolean, String, Float, etc.) has a corresponding native type.

fn native_type_is_default_for_scalar_type(
&self,
native_type: Value,
scalar_type: &ScalarType
) -> bool

Same mapping as default_native_type_for_scalar_type(), but in the opposite direction.

fn parse_native_type(
&self,
name: &str,
args: Vec<String>
) -> Result<NativeTypeInstance, ConnectorError>

This function is used during Schema parsing to calculate the concrete native type.

fn introspect_native_type(
&self,
native_type: Value
) -> Result<NativeTypeInstance, ConnectorError>

This function is used during introspection to turn an introspected native type into an instance that can be put into the Prisma schema.

fn validate_url(&self, url: &str) -> Result<(), String>

Provided methods

fn allowed_referential_integrity_settings(
&self
) -> BitFlags<ReferentialIntegrity>

The referential integrity modes that can be set through the referentialIntegrity datasource argument.

fn default_referential_integrity(&self) -> ReferentialIntegrity

The default referential integrity mode to assume for this connector.

fn has_capability(&self, capability: ConnectorCapability) -> bool

fn supports_composite_types(&self) -> bool

fn supports_named_primary_keys(&self) -> bool

fn supports_named_foreign_keys(&self) -> bool

fn supports_named_default_values(&self) -> bool

fn supports_referential_action(
&self,
integrity: &ReferentialIntegrity,
action: ReferentialAction
) -> bool

fn validate_field_default_without_native_type(
&self,
_field_name: &str,
_scalar_type: &ScalarType,
_default: Option<&DefaultValue>,
_errors: &mut Vec<ConnectorError>
)

fn validate_native_type_arguments(
&self,
_native_type: &NativeTypeInstance,
_scalar_type: &ScalarType,
_: &mut Vec<ConnectorError>
)

Validate that the arguments passed to a native type attribute are valid.

fn validate_model(&self, _: &Model, _: &mut Vec<ConnectorError>)

fn constraint_violation_scopes(&self) -> &'static [ConstraintScope]

The scopes in which a constraint name should be validated. If empty, doesn’t check for name clashes in the validation phase.

fn find_native_type_constructor(
&self,
name: &str
) -> Option<&NativeTypeConstructor>

fn render_native_type(&self, native_type: Value) -> String

This function is used in ME for error messages

fn set_config_dir<'a>(&self, config_dir: &Path, url: &'a str) -> Cow<'a, str>

fn supports_scalar_lists(&self) -> bool

fn supports_relations_over_non_unique_criteria(&self) -> bool

fn supports_enums(&self) -> bool

fn supports_json(&self) -> bool

fn supports_auto_increment(&self) -> bool

fn supports_non_id_auto_increment(&self) -> bool

fn supports_multiple_auto_increment(&self) -> bool

fn supports_non_indexed_auto_increment(&self) -> bool

fn supports_compound_ids(&self) -> bool

fn allows_relation_fields_in_arbitrary_order(&self) -> bool

fn native_instance_error(
&self,
instance: &NativeTypeInstance
) -> ConnectorErrorFactory

fn native_str_error(&self, native_str: &str) -> ConnectorErrorFactory

fn native_types_not_supported(
&self
) -> Result<NativeTypeInstance, ConnectorError>