Trait datamodel_connector::walker_ext_traits::ScalarFieldWalkerExt

pub trait ScalarFieldWalkerExt {
    fn native_type_instance(
        &self, 
        connector: &dyn Connector
    ) -> Option<NativeTypeInstance>;
}

Required methods

fn native_type_instance(
&self,
connector: &dyn Connector
) -> Option<NativeTypeInstance>

This will return None when:

There is no native type attribute on the field.

The native type attribute is not valid for the connector.