Trait datamodel_connector::Connector

pub trait Connector: Send + Sync {

    fn name(&self) -> &str;

    fn capabilities(&self) -> &'static [ConnectorCapability];

    fn max_identifier_length(&self) -> usize;

    fn referential_actions(
        &self, 
        referential_integrity: &ReferentialIntegrity
    ) -> BitFlags<ReferentialAction>;

    fn available_native_type_constructors(
        &self
    ) -> &'static [NativeTypeConstructor];

    fn scalar_type_for_native_type(&self, native_type: Value) -> ScalarType;

    fn default_native_type_for_scalar_type(
        &self, 
        scalar_type: &ScalarType
    ) -> Value;

    fn native_type_is_default_for_scalar_type(
        &self, 
        native_type: Value, 
        scalar_type: &ScalarType
    ) -> bool;

    fn parse_native_type(
        &self, 
        name: &str, 
        args: Vec<String>, 
        span: Span
    ) -> Result<NativeTypeInstance, DatamodelError>;

    fn introspect_native_type(&self, native_type: Value) -> NativeTypeInstance;

    fn validate_url(&self, url: &str) -> Result<(), String>;

    fn has_capability(&self, capability: ConnectorCapability) -> bool { ... }

    fn allowed_referential_integrity_settings(
        &self
    ) -> BitFlags<ReferentialIntegrity> { ... }

    fn default_referential_integrity(&self) -> ReferentialIntegrity { ... }

    fn supports_composite_types(&self) -> bool { ... }

    fn supports_named_primary_keys(&self) -> bool { ... }

    fn supports_named_foreign_keys(&self) -> bool { ... }

    fn supports_named_default_values(&self) -> bool { ... }

    fn supports_referential_action(
        &self, 
        integrity: &ReferentialIntegrity, 
        action: ReferentialAction
    ) -> bool { ... }

    fn validate_native_type_arguments(
        &self, 
        _native_type: &NativeTypeInstance, 
        _scalar_type: &ScalarType, 
        _span: Span, 
        _: &mut Diagnostics
    ) { ... }

    fn validate_model(&self, _model: ModelWalker<'_>, _: &mut Diagnostics) { ... }

    fn constraint_violation_scopes(&self) -> &'static [ConstraintScope] { ... }

    fn find_native_type_constructor(
        &self, 
        name: &str
    ) -> Option<&NativeTypeConstructor> { ... }

    fn set_config_dir<'a>(
        &self, 
        config_dir: &Path, 
        url: &'a str
    ) -> Cow<'a, str> { ... }

    fn supports_scalar_lists(&self) -> bool { ... }

    fn supports_relations_over_non_unique_criteria(&self) -> bool { ... }

    fn supports_enums(&self) -> bool { ... }

    fn supports_json(&self) -> bool { ... }

    fn supports_json_lists(&self) -> bool { ... }

    fn supports_auto_increment(&self) -> bool { ... }

    fn supports_non_id_auto_increment(&self) -> bool { ... }

    fn supports_multiple_auto_increment(&self) -> bool { ... }

    fn supports_non_indexed_auto_increment(&self) -> bool { ... }

    fn supports_compound_ids(&self) -> bool { ... }

    fn allows_relation_fields_in_arbitrary_order(&self) -> bool { ... }

    fn native_instance_error(
        &self, 
        instance: &NativeTypeInstance
    ) -> ConnectorErrorFactory { ... }

}

The datamodel connector API.

Required methods

fn name(&self) -> &str

The name of the connector. Can be used in error messages.

fn capabilities(&self) -> &'static [ConnectorCapability]

The static list of capabilities for the connector.

fn max_identifier_length(&self) -> usize

The maximum length of constraint names in bytes. Connectors without a limit should return usize::MAX.

fn referential_actions(
&self,
referential_integrity: &ReferentialIntegrity
) -> BitFlags<ReferentialAction>

The referential actions supported by the connector.

fn available_native_type_constructors(&self) -> &'static [NativeTypeConstructor]

Returns all available native type constructors available through this connector. Powers the auto completion of the VSCode plugin.

fn scalar_type_for_native_type(&self, native_type: Value) -> ScalarType

Returns the Scalar Type for the given native type

fn default_native_type_for_scalar_type(&self, scalar_type: &ScalarType) -> Value

On each connector, each built-in Prisma scalar type (Boolean, String, Float, etc.) has a corresponding native type.

fn native_type_is_default_for_scalar_type(
&self,
native_type: Value,
scalar_type: &ScalarType
) -> bool

Same mapping as default_native_type_for_scalar_type(), but in the opposite direction.

fn parse_native_type(
&self,
name: &str,
args: Vec<String>,
span: Span
) -> Result<NativeTypeInstance, DatamodelError>

This function is used during Schema parsing to calculate the concrete native type.

fn introspect_native_type(&self, native_type: Value) -> NativeTypeInstance

This function is used during introspection to turn an introspected native type into an instance that can be inserted into dml.

fn validate_url(&self, url: &str) -> Result<(), String>

Provided methods

fn has_capability(&self, capability: ConnectorCapability) -> bool

Does the connector have this capability?

fn allowed_referential_integrity_settings(
&self
) -> BitFlags<ReferentialIntegrity>

The referential integrity modes that can be set through the referentialIntegrity datasource argument.

fn default_referential_integrity(&self) -> ReferentialIntegrity

The default referential integrity mode to assume for this connector.

fn supports_composite_types(&self) -> bool

fn supports_named_primary_keys(&self) -> bool

fn supports_named_foreign_keys(&self) -> bool

fn supports_named_default_values(&self) -> bool

fn supports_referential_action(
&self,
integrity: &ReferentialIntegrity,
action: ReferentialAction
) -> bool

fn validate_native_type_arguments(
&self,
_native_type: &NativeTypeInstance,
_scalar_type: &ScalarType,
_span: Span,
_: &mut Diagnostics
)

Validate that the arguments passed to a native type attribute are valid.

fn validate_model(&self, _model: ModelWalker<'_>, _: &mut Diagnostics)

fn constraint_violation_scopes(&self) -> &'static [ConstraintScope]

The scopes in which a constraint name should be validated. If empty, doesn’t check for name clashes in the validation phase.

fn find_native_type_constructor(
&self,
name: &str
) -> Option<&NativeTypeConstructor>

fn set_config_dir<'a>(&self, config_dir: &Path, url: &'a str) -> Cow<'a, str>

fn supports_scalar_lists(&self) -> bool

fn supports_relations_over_non_unique_criteria(&self) -> bool

fn supports_enums(&self) -> bool

fn supports_json(&self) -> bool

fn supports_json_lists(&self) -> bool

fn supports_auto_increment(&self) -> bool

fn supports_non_id_auto_increment(&self) -> bool

fn supports_multiple_auto_increment(&self) -> bool

fn supports_non_indexed_auto_increment(&self) -> bool

fn supports_compound_ids(&self) -> bool

fn allows_relation_fields_in_arbitrary_order(&self) -> bool

fn native_instance_error(
&self,
instance: &NativeTypeInstance
) -> ConnectorErrorFactory